backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
sqlmodel>=0.0.16
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pydantic-settings>=2.2.1
openai>=1.0.0
cohere>=5.0.0
